+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## v8.40.0 (unreleased)
+## v8.40.0 (July 21st 2025)
 ### New features
 - [PR #2154](https://github.com/rqlite/rqlite/pull/2154), [PR #2156](https://github.com/rqlite/rqlite/pull/2156), [PR #2170](https://github.com/rqlite/rqlite/pull/2170): Add a HTTP API which instructs the Leader to stepdown, transferring Leadership to another node in the cluster. Fixes issue [#2153](https://github.com/rqlite/rqlite/issues/2153).
 - [PR #2159](https://github.com/rqlite/rqlite/pull/2159): Add `.stepdown` command to rqlite shell.